added handlers methods to some classes
