+++ src/utils/render.js
@@ -1,0 +1,16 @@
+export const createElement = (template) => {
+  const newElement = document.createElement(`div`);
+  newElement.innerHTML = template;
+  return newElement.firstChild;
+};
+
+export const render = (container, node, place = `beforeend`) => {
+  switch (place) {
+    case `afterbegin`:
+      container.prepend(node);
+      break;
+    case `beforeend`:
+      container.append(node);
+      break;
+  }
+};
